# api/admin-api-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing info about allowed services and targets to proxy",
        "title": "Proxy Service Admin API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/admin",
    "paths": {
        "/ping": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common API"
                ],
                "summary": "Ping",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.StatusResponse"
                        }
                    }
                }
            }
        },
        "/v1/service": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "Get all services with targets",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/proxy-service_internal_models.ServiceDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "Create service",
                "parameters": [
                    {
                        "description": "Service data (with targets)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/proxy-service_internal_models.ServiceDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Service already exists",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/service/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "Get service with targets by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/proxy-service_internal_models.ServiceDTO"
                        }
                    },
                    "404": {
                        "description": "Service not found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "Update service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Service data (with targets)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/proxy-service_internal_models.ServiceDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Service not found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "Delete service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.StatusResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "proxy-service_internal_models.ServiceDTO": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "retry_count": {
                    "type": "integer"
                },
                "retry_interval": {
                    "type": "number"
                },
                "scheme": {
                    "type": "string"
                },
                "targets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/proxy-service_internal_models.TargetDTO"
                    }
                },
                "timeout": {
                    "type": "number"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "proxy-service_internal_models.TargetDTO": {
            "type": "object",
            "properties": {
                "cache_interval": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                }
            }
        }
    }
}